Allow all branches to push
gh-7

Still trying to get rid of the error:

Error: Not authorized to perform sts:AssumeRoleWithWebIdentity

I can see example repos that work, and the only apparent difference is
they have a wildcard as is added here.

This is not an acceptable long-term solution, because I want to use
GitHub's branch control rules to restrict who can deploy to the account
associated with the master branch.